refactor: LiveActivity を独立パッケージ化、Monitor/Manager の重複を解消
- 新規パッケージ Packages/LiveActivity
  - TodoLiveActivityManager: ActivityKit の start/end/update ラッパー
  - LiveActivityMonitor: Todo 配列監視 + Manager 委譲の SwiftUI modifier
  - iOS 限定 API のため全ソース #if os(iOS) ガード
- UI/Sources/UI/LiveActivity/LiveActivityMonitor.swift を削除（新パッケージに移送）
- IntentTodoLiveActivity/Manager/TodoLiveActivityManager.swift を削除（dead code だったのを本パッケージに昇格）
- UI パッケージの Package.swift に LiveActivity 依存を追加
- TodoListView は import LiveActivity して .monitorLiveActivities を使用
- project.pbxproj: LiveActivity フォルダ参照追加 + 旧 Manager ファイル関連4箇所削除
