+++ Packages/UI/Sources/UI/Views/TodoList/TodoListView.swift
@@ -4,6 +4,9 @@
 //
 
 import Domain
+#if os(iOS)
+import LiveActivity
+#endif
 import SwiftData
 import SwiftUI
 import TodoAppIntents
